Rewrite make(BioEntity, SequenceType) to actually use the SequenceType, not just Protein=amino acids, Else=DNA
This method seems to have been mis-written originally. It did not use the SequenceType type parameter at all. It simply checked if the BioEntity is a Protein, in which case it returned an amino acid sequence, else it assumed the BioEntity was a DNA sequence. However, we have Polypeptides imported from chado as SequenceFeatures with amino acid residues. So I have rewritten this method to do what I think it is meant to do: use the input SequenceType to determine the type of sequence. (Ideally, SequenceFeature and Protein would have a standard field specifying the SequenceType and the three variations of make() would not be needed. There are sequence features like polypeptides in our case that are associated with a DNA locus but have PROTEIN or RNA residues.)

Former-commit-id: e19a9c364387035c3f5f3840f0ae48086f845e95