Update GenomicRegionSearchService.java
Bug fix - the logic building errorSpanList was faulty, since regions added to passedSpanList after their coordinates are reversed are NOT members of grsc.getGenomicRegionList(). Therefore, all reversed-coordinate (- strand) regions were added to errorSpanList. New logic simply uses a flag "passed" that is set if the region is added to passedSpanList; otherwise, it is added to errorSpanList.

Former-commit-id: 01c28af40b0e13d4372f3afffe31bd0e29d58cbb